169664175: Arrumando bug de migration
