+++ .env.example
@@ -14,18 +14,11 @@ PROJECT_NAME=rodamais
 
 # NODEJS
 
-ENVIRONMENT=dev 
+ENVIRONMENT=dev
 
 # TYPEORM
-TYPEORM_CONNECTION=postgres
-TYPEORM_HOST=postgres
-TYPEORM_USERNAME=admin
-TYPEORM_PASSWORD=rodamais
-TYPEORM_DATABASE=rodamais
-TYPEORM_PORT=5432
-TYPEORM_SYNCHRONIZE=true
-TYPEORM_LOGGING=true
-TYPEORM_ENTITIES=dist/**/**.entity.js
-TYPEORM_MIGRATIONS=migration/*.ts
-TYPEORM_SEEDS=seeds/**/*.seed.ts
-TYPEORM_FACTORIES=factories/**/*.factory.ts
+DB_USER=admin
+DB_PASSWORD=rodamais
+DB_DATABASE=rodamais
+DB_PORT=5432
+DB_HOST=postgres
