+++ docker-compose.yml
@@ -10,9 +10,9 @@ services:
     ports:
       - 5432:5432
     environment:
-      POSTGRES_USER: ${TYPEORM_USERNAME}
-      POSTGRES_PASSWORD: ${TYPEORM_PASSWORD}
-      POSTGRES_DB: ${TYPEORM_DATABASE}
+      POSTGRES_USER: ${DB_USER}
+      POSTGRES_PASSWORD: ${DB_PASSWORD}
+      POSTGRES_DB: ${DB_DATABASE}
     volumes:
      - ${DATA_SAVE_PATH}/postgres:/var/lib/postgres
     networks:
