+++ ormconfig.json
@@ -1,21 +1,20 @@
 {
   "type": "postgres",
   "host": "postgres",
-  "port": "5432",
+  "port": 5432,
   "username": "admin",
   "password": "rodamais",
   "database": "rodamais",
   "entities": [
-    "dist/src/**/*.entity.js",
-    "src/**/*.entity.ts"
+    "dist/**/*.entity.js"
   ],
   "migrations": [
-    "migration/*{.ts,.js}"
+    "migration/*.js"
   ],
   "cli": {
     "migrationsDir": "migration"
   },
-  "synchronize": false,
+  "synchronize": true,
   "logging": [
     "error",
     "warn",
