+++ package.json
@@ -7,7 +7,7 @@
   "scripts": {
     "ts-node": "ts-node",
     "build": "rimraf dist && tsc -p tsconfig.build.json",
-    "seed": "ts-node ./node_modules/typeorm-seeding/dist/cli.js seed --config ormconfig.json",
+    "seed": "npm run typeorm schema:drop && npm run typeorm migration:run && ts-node ./node_modules/typeorm-seeding/dist/cli.js seed --config ormconfig.json",
     "format": "prettier --write \"src/**/*.ts\"",
     "start": "ts-node -r tsconfig-paths/register src/main.ts",
     "start:dev": "rimraf dist && tsc-watch -p tsconfig.build.json --onSuccess \"node dist/main.js\"",
