+++ tslint.json
@@ -1,21 +1,50 @@
 {
-  "defaultSeverity": "error",
-  "extends": ["tslint:recommended"],
-  "jsRules": {
-    "no-unused-expression": true
-  },
+  "extends": [
+    "tslint-eslint-rules"
+  ],
+  "rulesDirectory": [],
   "rules": {
-    "no-console": false,
-    "quotemark": [true, "single"],
-    "member-access": [false],
-    "ordered-imports": [false],
-    "max-line-length": [true, 150],
-    "member-ordering": [false],
-    "interface-name": [false],
+    "eofline": false,
+    "trailing-comma": false,
+    "member-access": false,
+    "indent": false,
+    "ordered-imports": [
+      false
+    ],
+    "no-implicit-dependencies": false,
+    "no-unnecessary-initializer": false,
+    "unified-signatures": false,
+    "callable-types": false,
+    "max-line-length": [
+      true,
+      150
+    ],
+    "member-ordering": [
+      false
+    ],
+    "curly": false,
+    "interface-name": [
+      false
+    ],
+    "array-type": [
+      false
+    ],
+    "no-empty-interface": false,
+    "no-empty": false,
     "arrow-parens": false,
     "object-literal-sort-keys": false,
-    "no-empty": false,
-    "curly": false
-  },
-  "rulesDirectory": []
+    "no-unused-expression": false,
+    "max-classes-per-file": false,
+    "ban-types": false,
+    "variable-name": [
+      false
+    ],
+    "one-line": [
+      false
+    ],
+    "one-variable-per-declaration": [
+      false
+    ],
+    "prefer-readonly": true
+  }
 }
